build_image: Include a list of models in unified builds
When unified builds we want to include a list of models for display on the
login screen if enabled. Update the

BUG=chromium:698012
TEST=Manual:
USE=pcserial ./build_image --board=reef --noenable_rootfs_verification dev \
   --enable_serial ttyS0 dev
Boot it on reef in dev mode and see that the message is displayed on the
login screen.

Change-Id: I4017dd30092eca406985034235ce23ddd0907f44
Reviewed-on: https://chromium-review.googlesource.com/450783
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>