package_to_container: set up for imageloader
This changes the package setup so that we store the directory
structure expected by run_oci entirely in the squashfs image, and
then set up the rest of the container to be mounted with
imageloader.

Since the rootfs is now contained in the squashfs image instead
of mounting the squashfs image into the rootfs directory, we
also get rid of the root mount point entry in the config.json
template for run_oci.

BUG=chromium:697645
TEST=package adb container, mount with imageloader using dev keys,
  run_oci --unsigned the resulting mount point (or using command
  introduced in CL:362122)

Change-Id: I9eb40c4865a462e702dbbd9e48ea64dbcb822aa0
Reviewed-on: https://chromium-review.googlesource.com/456924
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>