Add new arm_trusted_firmware_uprev script
After long discussion we have decided to change the way the ARM Trusted
Firmware repository is handled from manifest pinning to a master branch
that is fast-forwarded directly to new upstream commits through direct
git pushes. This is supposed to improve trybot compatibility, among
other things.

This script is intended to make that fast-forwarding as brain-dead
simple as possible for all the firmware developers who'll have to do it
regularly. It will automatically find the new upstream commit and push
the appropriate branches. It also includes an automated system of saving
older known-good HEADs in branches named prev1 through prev5 to make it
easy for sheriffs to revert a bad push without rewriting history if
necessary (by changing the manifest to temporarily point to a prev
branch).

BUG=chromium:630016
TEST=Run script, get git push

Change-Id: I29b3ed5807d56c761fcbe38b1e23105a671dacc5
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/497126
Reviewed-by: Mike Frysinger <vapier@chromium.org>