Add a script to automate firmware update process
A typical process to uprev firmware is:
(1) Download a specific firmware tarball from GE or CPFE
(2) Untar the tarball and then repack new tarballs for AP/EC firmware
    binaries respectively
(3) Upload the new tarballs to BCS
(4) Modify the firmware ebuild in device private overlay
(5) Build firmware updater and verify the firmware version is correct
(6) (Optional) Test the firmware updater on a real device
(7) Submit a CL and wait for approval

This script aims to automate 1-2 & 4-5.
(3) needs to be done by CPFE manually because of the limitation of access.
(7) needs more customization by the user.

BUG=None
TEST=uprev firmware for bob and gru

Change-Id: Ia3f111c79af03a45139aad2618e4724da5ed91ec
Reviewed-on: https://chromium-review.googlesource.com/477471
Commit-Ready: Philip Chen <philipchen@chromium.org>
Tested-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>