Switch default SDK compiler to clang.
This CL is part of the migration of the SDK default compiler from
GCC to clang.  It updates the function that generates
/etc/make.conf.host_setup to allow it to specify whether or not to force
CC/CXX to be clang.  If the argument is passed to do this, then it appends
the appropriate CC=... and CXX=... to the end of make.conf.host_setup.
It also updates the places that call the function, to pass the appropriate
arguments.

BUG=chromium:674175
CQ-DEPEND=CL:482475
TEST=Currently testing with various buildbots.

Change-Id: I2df399ecca9bec38cff9a16596d26fb8227360f4
Reviewed-on: https://chromium-review.googlesource.com/481864
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>