uprev_firmware.sh: Fix a bug in rw_only firmware update
A recent change in firmware updater suggests that we define
'CROS_FIRMWARE_MAIN_RW_IMAGE' as an empty string early in the
firmware build.

But this script won't work well for rw_only firmware update
when the empty string of 'CROS_FIRMWARE_MAIN_RW_IMAGE'
is already defined in the ebuild.

Here is the fix.

BUG=chromium:705414
TEST=uprev_firmware -b bob -v 8785.202.0 rw_only

Change-Id: Ic6d0e588f60b45b786308e52551dc45c975f7898
Reviewed-on: https://chromium-review.googlesource.com/510210
Commit-Ready: Philip Chen <philipchen@chromium.org>
Tested-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>