update_kernel: error out if kernel partition isn't found
The kernel partition might not be found for a number of reasons:

 * the '--partition=' argument might be incorrect
 * the DUT might not be compatible with us (e.g., it's using an old
   image, with differences in /usr/sbin/write_gpt.sh)

Currently, rather than noticing this and bailing out, we're happy to
use 'dd' to create a new file instead. That can create unintuitive
behaviors, where the DUT's kernel modules have been updated while the
kernel partition is left unmodified.

Let's detect this situation and error out, instead of leaving the user
to figure out what happened.

BUG=chromium:733028
TEST=update_kernel on old images (without $PARTITION_NUM... variables)
     and new ones

Change-Id: I47f64ec7fb1175e0e25697a6b55b63366fcc4555
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/537140
Reviewed-by: Mike Frysinger <vapier@chromium.org>