disk layout: raise default state filesystem size to 2 GiB
We've been (close to) filling the stateful for a while now due to all the
test data we throw into it, so make it bigger.  This shouldn't impact AU
at all as we only process the rootfs & kernel, and we're only increasing
the filesystem size, not the partition size.

BUG=chromium:753498
TEST=precq passes

Change-Id: Ic3fa9d63607172c196d4bce76e8a949e0729f044
Reviewed-on: https://chromium-review.googlesource.com/611724
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jacob Dufault <jdufault@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>