setup_board: automatically include sdk toolchain packages in host boards
Rather than requiring the list of toolchain packages that we are managing
by hand be listed in the virtual/target-sdk tree, pull in the list when
we are creating a new host board from scratch.  This shouldn't currently
be functionally different, but should smooth over future changes since
there will be only one place to maintain this list.

BUG=chromium:751852
TEST=precqs pass
CQ-DEPEND=CL:664840

Change-Id: Ia54472b80efd8c3fa7695d42df10027e1470105f
Reviewed-on: https://chromium-review.googlesource.com/666297
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>