setup_board: automatically include sdk toolchain packages in host boards [reland]
Rather than requiring the list of toolchain packages that we are managing
by hand be listed in the virtual/target-sdk tree, pull in the list when
we are creating a new host board from scratch.  This shouldn't currently
be functionally different, but should smooth over future changes since
there will be only one place to maintain this list.

This version fixes the array usage.

BUG=chromium:751852
TEST=precqs pass

Change-Id: Ibc901dd7b45d1deacefdd6c6436fc53c4fcd8991
Reviewed-on: https://chromium-review.googlesource.com/669440
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>