remote_access: don't dump all processes, backtrace when reboot fails
'die()' includes a dump of all processes in an ASCII pretty-print, as
well as a call-stack of the current shell script. Neither is very useful
when the remote DUT doesn't come back from reboot properly (e.g., with
'update_kernel.sh').

This is especially bad when I'm running my chroot without a PID
namespace, and so I get >100 lines of useless garbage from 'ps'.

Let's use the notrace version, to get a much nicer output.

BUG=none
TEST=`update_kernel.sh ...`, where DUT doesn't come back

Change-Id: Ib114a96c68426b6dc5b881e9a708c81f8ee52361
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/690580
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>