common.sh: Add android build directory to COMMON_INSTALL_MASK
/opt/containers/google/android/build contains files that are needed to
build various arc-* packages.  These files are not needed at runtime and
shouldn't end up on the board.  In the past, we've lucked out because
arc-build isn't a dependency of virtual/target-os, but the new arc-llvm
package gets (correctly) pulled in as a runtime dependency.

Mask out the build directory so that these build-time files are never
included in the image.

BUG=b:65414758
TEST=sudo /mnt/host/source/chromite/bin/test_image \
     --board kahlee \
     ~/trunk/src/build/images/kahlee/latest/chromiumos_test_image.bin

Change-Id: Ic1d167abe2acfa4708002be77943a3cb4f6e0944
Reviewed-on: https://chromium-review.googlesource.com/736019
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>