mount_gpt_image: move secondary buildroot to /build
CL:780265 added a local buildroot under the sysroot that ebuilds can use
to install temporary files when building the image.
That breaks the devinstall test since it tries to wipe dev_image and
can't because the local buildroot is busy (it's a bindmount from a
different directory) and is under dev_image (in /usr/local/build).

Change the path of the secondary buildroot to /build to avoid
interfering with the stateful partition. Also, only create the secondary
buildroot when the root partition is RW, i.e. when neither --read_only
nor --safe flags have been enabled.

BUG=b:67717927
TEST=devinstall_test.py now passes.
CQ-DEPEND=CL:*527462

Change-Id: I8e4530904f94221722c90262659737a28d533989
Reviewed-on: https://chromium-review.googlesource.com/822529
Commit-Ready: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>