cros_vm_lib: Add option to configure qemu serial output
For debugging purposes and for images without graphics, an interactive
serial console is useful for communication with the guest VM. Add the
flag --serial, whose argument is used as the '-serial' argument to qemu.
For example, to have qemu allocate a PTY for serial console access, pass
"--serial=pty". By default, serial console output is written to a file.

BUG=none
TEST='bin/cros_start_vm --image_path=... --no_graphics --serial=pty' and
  connect to allocated PTY with screen.

Change-Id: I779385bf25a4162bfd3f0247eafb353f1c897604
Signed-off-by: Andrew Bresticker <abrestic@waymo.com>
Reviewed-on: https://chromium-review.googlesource.com/1062445
Commit-Ready: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>