arm_trusted_firmware_uprev: Clarify error message for most common case
Direct Git push access must always be restricted, so the
arm_trusted_firmware_uprev script only works for people who're part of
the chromeos-arm-trusted-firmware-pushers Ganpati group. This is easy to
forget when telling new developers who just started to work on Arm TF
about the script. This patch adds a more explicit error message to this
case to help point them in the right direction.

BUG=none
TEST=none

Change-Id: I2c1daeb0ef281f7d666f09069dd1561f2a17e311
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1103387
Reviewed-by: Hung-Te Lin <hungte@chromium.org>