clean_loopback_devices: punt
This is an ancient script dating back to the start of the project.
Nothing in our build runs it, and it looks like it was meant to be
run manually by developers.  We've since vastly improved our usage
of loopback devices, so there shouldn't generally be leaking issues
anymore that people have to manually clean up.

BUG=None
TEST=precq passes

Change-Id: Id056d13e79fd722426ac6bb224627aedda3aadd6
Reviewed-on: https://chromium-review.googlesource.com/1138804
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>