chroot_version_hooks.d: drop support for upgrading SDKs pre-2015
Punt a lot of old upgrade hooks that only get used for SDKs that were
created before 2015 and have not been upgraded since.  The tooling will
already detect this scenario (because the upgrade hooks are missing) and
throw an error telling the user to recreate the SDK.  Chances are good
that it wouldn't actually work, and we certainly have no testing to make
sure that it does.  So punt it all.

BUG=None
TEST=precq passes

Change-Id: I99323c13919ff4825b9f0120729ae64e3950ca34
Reviewed-on: https://chromium-review.googlesource.com/1163170
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>