build_packages with fewer revdeps
Cleaning up reverse dependency calculations. Removing saving
the install plan currently calculated in parallel_emerge.py
with a grep of the emerge simulation output. This will only
pickup the changes that are to ebuilds instead of the previously
incorrect reverse dependencies on the binary packages.

Will cleanup the rest of the install plan code in another CL once
this CL is working correctly locally and in the CQ.

BUG=chromium:864309
TEST=local builds

Change-Id: Id77ad050679fb701d2b455e462fc91ba1c09d0ff
Reviewed-on: https://chromium-review.googlesource.com/1187194
Commit-Queue: Gregory Meinke <gmeinke@chromium.org>
Tested-by: Gregory Meinke <gmeinke@chromium.org>
Trybot-Ready: Gregory Meinke <gmeinke@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gregory Meinke <gmeinke@chromium.org>