setup_board: Use virtual/target-sdk-nobdeps on host
virtual/target-sdk-nobdeps pulls in ebuilds that should be installed
without their build-deps.  This gets built as normal during the first
pass of rebuilding the chroot, then gets pulled into the new board root
with --with-bdeps=n to ensure that build-time dependencies are not
present in the final chroot image.  The initial use case is to make
shellcheck available in the chroot without bringing in all the Haskell
compiler dependencies.

BUG=chromium:844166
TEST=chromiumos-sdk-tryjob
CQ-DEPEND=CL:1070324

Change-Id: I536d7da844ecd29a09e4b3a7fb8e922328bcb465
Reviewed-on: https://chromium-review.googlesource.com/1260045
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>