make_chroot: manually build host toolchain packages in order
To workaround race conditions where we build & install a package that
is actively being used by other packages we're building (e.g. we run
the assembler while building gcc while also building & installing the
assembler in parallel), pull some of the critical packages out so we
can run them in a specific order.

For now we build binutils & the C library in serial.  The other libs
and compilers don't directly depend on each other, so they can still
build in parallel.

This might slow the build down overall slightly, but we probably don't
run as slow in practice because we don't see large packages (e.g. gcc)
failing to build and then needing to be retried.

BUG=chromium:715788
TEST=bootstrapping sdk still works

Change-Id: I5d52d1660662957daf10f4ba6b0d116a788dff21
Reviewed-on: https://chromium-review.googlesource.com/c/1387715
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>