build_sdk_board: build the toolchain packages by themselves
By building the toolchain in parallel with other packages which rely
on the toolchain, we might introduce random breakage when a file is
updated in place and temporarily causes it to fail to run (e.g. if
we update the assembler while it's running, it might fail).  We can
mitigate this by building just the toolchain packages by themselves
followed by everything else.

BUG=chromium:906289
TEST=sdk bot passes

Change-Id: I1c8fcf74f232b66de52def84d0ce2f05b0bc16b0
Reviewed-on: https://chromium-review.googlesource.com/c/1387717
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>