build_kernel_image: Don't specify oops=panic and panic=-1
To avoid specifying the same thing in lots of different places, we're
moving to setup the panic behavior in only one place: in the kernel
config.  Let's remove it from the kernel command line.

See the linked bug (chromium:908726) for all the CLs that add this to
all of our kernel configs.  In case someone needs to do this on some
private kernel, they could do:
  echo 'CONFIG_PANIC_TIMEOUT=-1' >> chromeos/config/base.config
  echo 'CONFIG_PANIC_ON_OOPS=y' >> chromeos/config/base.config

NOTE: for older kernels (known to be needed in 3.10 and older) you'll
need an extra cherry-pick for the above config options to work.
Presumably this is why we originally did this on the command line.
Again, see the linked bug.

BUG=chromium:908726
TEST=tast -verbose run ${IP} kernel.Limits, which is part of the CQ.

Change-Id: Ic008cfe6f1493575f57084d562ab545738524912
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1401477
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>