build_image: move package.installable to $SYSROOT/build/dev-install
This file is only read by the builders in chromite for uploading files,
so it doesn't need to be in the rootfs.  Move it over to the /build/
subdir inside of the sysroot.

We also fix the generation of this file.  chromite was using the file
as it existed in the sysroot (pre-build_image), so the problems in this
file weren't noticed.  We need to remove duplicates from the two input
lists otherwise the output contains packages it shouldn't (such as ones
installed in the rootfs already).

BUG=chromium:202477
TEST=precq passes

Change-Id: Ib0a4c59132ccfccb0735a08b416a40009cdfe578
Reviewed-on: https://chromium-review.googlesource.com/1484464
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>