dev-install: move package.provided out of the default profile
We want this file when we have a base image and we've run `dev_install`.
If we have a dev/test image, we don't need this as we'll have the full
/var/db/pkg environment.  So lets move it to /usr/share/dev-install so
the dev_install tool can manually set it up.  We create a new subdir at
/usr/share/dev-install/rootfs.provided/ and dev_install will install all
files that are in there.

We leave the old entry in place for now to help with merging CLs.

BUG=chromium:920140
TEST=precq passes

Change-Id: I581bc1843b1d6cb14e2bf9adcdf89c03632450d8
Reviewed-on: https://chromium-review.googlesource.com/1494234
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>