build_image: move /usr/share/portage symlink to base images
For dev images, we copy portage files from /usr/local to /usr so we
can run portage in dev images.  This is because portage hardcodes the
/usr path in its source files regardless of where it's installed.
Instead of doing this for just dev images, lets add a symlink into
the base image so portage works with dev-install too.

BUG=chromium:935142
TEST=precq passes
CQ-DEPEND=CL:1484754

Change-Id: Iaf14f8bd4b24cc150af61a929613942c45c27ef9
Reviewed-on: https://chromium-review.googlesource.com/1495194
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>