build_image: copy over profile provided settings
This setting is currently handled by the gmerge package.  We want to
drop that and let it be part of the build_image stage.  These packages
are never built for the board, so fixing it at build_image time should
be OK.

BUG=chromium:194889
TEST=precq passes & dev_install works

Change-Id: I5e7613561f71b4bb475aa04be70ca56ab8a707c9
Reviewed-on: https://chromium-review.googlesource.com/1490620
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>