update_kernel: Add boot_command flag to run commands on remote after update
Allows the caller to specify a command to run on the remote after the
kernel has updated (and rebooted). Useful for things that either can't
be specified on the command line, or for times when you don't want to
mess with the command line.

Change-Id: Ie975105ac148c6c7ceea4c34895408133d7830b3
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1529148
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>