termina_build_image: fix file system shrink.
Warn users to install fdupes when it's not available.
Also, resize2fs doesn't work for me on 1K block size and man page of resize2fs
reads "The minimum size of the filesystem as estimated by resize2fs may be
incorrect,  especially for filesystems with 1k and 2k blocksizes.", so just
begin with the 110% of disk usage of root fs.

BUG=None
TEST=manual - run termina_build_image

Change-Id: I54c8a6e342f6e6d4c1ed5c0c27d72f81b19f9013
Reviewed-on: https://chromium-review.googlesource.com/1544864
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Lepton Wu <lepton@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>