cgpt: Fix fs_block_align_losses for fs_align
cgpt tries to account for possible alignment losses when figuring out
the total size of the image. It does this by adding a fixed amount to
each partition, and hoping that's enough. It turns out to be enough
only for the default value of fs_block_size.

Use fs_align for file system partitions, which may align up to a much larger
value, and fs_block_size for all other partitions.

BUG=b:130052781
TEST=build_image --board=hatch (and kukui, cheza, kalista)

Change-Id: I16e16ce1edf6b326fcd98bf222a367a5951b109f
Reviewed-on: https://chromium-review.googlesource.com/1555516
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>