update_kernel: Learn arch from board build (redux)
The boot.desc file in a build contains the information about what arch
the build is for. But sometimes that file isn't built, e.g. when a dev
has only built the kernel and not a whole image. Parse the kernel config
file for the architecture instead and pass it along to vbutil. This
avoids SSH to the DUT just to run uname.

BUG=None
TEST=update kernel on Cheza works

Change-Id: I7263ca68b5fd9ab1559630a8e51dd6c93c0155c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/1907980
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>