enter_chroot: check for chrome cache dir inside the chroot
Since the gclient path is going to be interpreted inside the chroot,
see if it exists inside of there.  We've assumed that the checkout
was created & fully used outside of the SDK, but if it's managed in
the SDK, requiring paths to also exist outside of the SDK doesn't
make sense.

BUG=chromium:991603
TEST=`cros_sdk` doesn't warn

Change-Id: I55c78104b4af0b5ae7763e330e23c233cc8ab014
Reviewed-on: https://chromium-review.googlesource.com/1742537
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>