termina_build_image: transition to Python version
The shipping Crostini Termina images are built by a reimplementation of
termina_build_image in Python, which lives in the container-guest-tools
repository. The Python version also works outside of the cros_sdk chroot
environment.

In order to reduce duplication and minimize confusion, replace the bash
implementation with a transitional script that calls the Python version.

BUG=chromium:999698
TEST=./termina_build_image --image=chromiumos_test_image.bin --output=/tmp/termina

Change-Id: Ibac7c5963892e76f4a606540e9943a057c26e529
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1779154
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>