drop show_help_if_requested
We added this logic when we were trying to make these programs more
"user friendly" and hiding options that were for builders only.  At
this point, only three scripts were using it anymore (build_sdk_board
is only for builders).  Lets drop this helper to tighten up logic a
bit even if it means these scripts (re)gain user-visible options.

When these move to chromite, we have more flexibility to mark options
as "builder" or "advanced".  We did that with setup_board already.

BUG=None
TEST=CQ passes

Change-Id: I5552322eb5d6f6209a017f01f65455a46eabc907
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/1838453
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>