build_image: fix dev-install package list w/bdepends
The new bdepends logic will include portage related packages in the
output of virtual/target-os.  The code here never expected that which
caused the bootstrap list to always be empty.  Make sure we filter
out all bdepend packages, and improve the emerge command a bit with
the --cols option so our parser doesn't need to be as complicated.

BUG=chromium:1015253
TEST=CQ passes

Change-Id: Ie1f6563b38170688644ea48bcec69d1aea941a55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/1885122
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>