bash_completion: Fix -- flags detection
This fixes the help flags detection regex.
The previous regex expected flags to have a colon trailing the flag
name. This resulted in no flags ever being reported by the _flags func.
This colon does not appear to exist anymore in the --help menus.

This fix changes the regex to look for -- flags that contain
alpha numeric characters, =, -, or _s.

BUG=none
TEST=cros_workon --<tab><tab>
TEST=cros_workon start --<tab><tab>
TEST=setup_board --<tab><tab>

Change-Id: Ib081ff93889c65ff750e34cad6f9abe07046afdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/1893497
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>