build_packages: Add 'use_any_chrome' flag
This flag directs the build to use any chrome prebuilt that is available
even if the USE flags are not an exact match.

Developers will have to manually re-emerge the stable version of
chromeos-chrome if they have stopped cros-workon for chromeos-chrome and
want to re-run build_packages using the stable chrome ebuild.

BUG=chromium:1015565
TEST=`build_packages` uses the chrome_internal prebuilt for veyron_minnie
TEST=`build_packages` with chromeos-chrome workon'd still builds the 9999 version

Change-Id: Id120189a73742f06efe6dacf34cd3e6c36090e7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/1913304
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>