update_kernel: Remove depmod logic
All of this 'run depmod' logic is unnecessary. Originally we introduced
this logic in here because chromium:451798 was making up for the fact
that the eclass change (chromium:446890) no longer installed kernel
modules. But that broke the waterfall and the eclass change got reverted
(see chromium:455341). In the end, we were left with running depmod in
update_kernel.sh when we don't really need to because the kernel
makefile will run depmod for us when 'modules_install' is called by the
kernel eclass in the src_install step.

Let's just remove this logic here to avoid it all.

BUG=chromium:695675
TEST=update_kernel on cheza, see that modules.dep matches
/build/cheza/lib/modules/<version>/modules.dep

Change-Id: Ic9c1e58f704a01789550e9fa7f628eb9f89c1e41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/1933594
Tested-by: Stephen Boyd <swboyd@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>